+++ src/util.ts
@@ -1,0 +1,150 @@
+import * as fs from "fs"
+import * as path from "path"
+import slash from "slash"
+import isDotfile from 'is-dotfile'
+import isDotdir from 'is-dotdir'
+import { AssembledOptions } from "./types"
+
+const shouldSkip = (relativeDir: string, skipDirectories: Array<string>) => {
+    for (let i = 0; i < skipDirectories.length; i += 1) {
+        if (relativeDir.startsWith(skipDirectories[i])) {
+            return true
+        }
+    }
+
+    return false
+}
+
+const isBlacklisted = (relativeDir: string, filename: string, options: AssembledOptions) => {
+    const {
+        excludeDirectories,
+        excludeFilenames,
+    } = options
+
+    // Check for blacklisted directories
+    for (let i = 0; i < excludeDirectories.length; i += 1) {
+        if (relativeDir.startsWith(excludeDirectories[i])) {
+            return true
+        }
+    }
+
+    // Check for blacklisted filename matches
+    for (let i = 0; i < excludeFilenames.length; i += 1) {
+        if (filename.match(excludeFilenames[i])) {
+            return true
+        }
+    }
+
+    return false
+}
+
+export const getFsEntity = (fullPath: string) => {
+    let dirent = null
+
+    try {
+        dirent = fs.statSync(fullPath)
+    } catch (e) {
+        // Ingore the error
+        // Scenarios this catches: symlinks or otherwise inaccessible directories
+    }
+
+    return dirent
+}
+
+export const flattenEntireTree = (items) => {
+    let flattened = []
+    items.forEach((item) => {
+        flattened.push(item)
+
+        if (item.type === 'dir') {
+            flattened = flattened.concat(flattenEntireTree(item.files))
+        }
+    })
+
+    return flattened
+}
+
+// Should the dir/file show at all in results?
+export const shouldSeeEntity = (dir: string, item: string, options: AssembledOptions) => {
+    const fullPath = path.join(dir, item)
+    const entity = getFsEntity(fullPath)
+    const relativeDir = slash(dir).replace(slash(process.cwd()), '')
+
+    // Is a symlink, or other un-readable entity?
+    if (!entity) {
+        return false
+    }
+
+    // Is a directory?
+    if (entity.isDirectory()) {
+        // Exclude dot directories, allow all others
+        const dotDir = isDotdir(fullPath) || isDotdir(slash(fullPath))
+        const dotFile = isDotfile(fullPath) || isDotfile(slash(fullPath))
+        if (dotDir || dotFile) {
+            return false
+        }
+
+        return true
+    }
+
+    // Is a dotfile?
+    const dotFile = isDotfile(fullPath) || isDotfile(slash(fullPath))
+    if (dotFile) {
+        return false
+    }
+
+    // Is codehawk config?
+    if (item === 'codehawk.json') {
+        return false
+    }
+
+    // Skip directories according to config
+    if (shouldSkip(relativeDir, options.skipDirectories)) {
+        return false
+    }
+
+    return true
+}
+
+// Should the dir/file have complexity analysis run?
+export const shouldAnalyzeEntity = (dir: string, item: string, options: AssembledOptions) => {
+    const fullPath = path.join(dir, item)
+    const entity = getFsEntity(fullPath)
+    const filename = path.basename(fullPath)
+    const extension = path.extname(filename)
+    const relativeDir = slash(dir).replace(slash(process.cwd()), '')
+
+    // Is a symlink, or other un-readable entity?
+    if (!entity) {
+        return false
+    }
+
+    // Is a directory?
+    if (entity.isDirectory()) {
+        // Exclude dot directories, allow all others
+        const dotDir = isDotdir(fullPath)
+        const dotFile = isDotfile(fullPath) || isDotfile(slash(fullPath))
+        if (dotDir || dotFile) {
+            return false
+        }
+
+        return true
+    }
+
+    const dotFile = isDotfile(fullPath) || isDotfile(slash(fullPath))
+    if (dotFile) {
+        return false
+    }
+
+    const isWhitelisted = options.extensions.includes(extension)
+    if (!isWhitelisted) {
+        return false
+    }
+
+    const blacklisted = isBlacklisted(relativeDir, filename, options)
+    if (blacklisted) {
+        return false
+    }
+
+    return true
+}
