+++ test/codehawk.js
@@ -1,6 +1,6 @@
 const { assert } = require('chai')
 const fs = require('fs')
-const codehawk = require('../src/codehawk')
+const codehawk = require('../build/codehawk.js')
 
 describe('codehawk.analyzeProject', () => {
     const cwd = process.cwd()
@@ -15,43 +15,43 @@ describe('codehawk.analyzeProject', () => {
         assert.deepEqual(output.results, expected.results)
     })
 
-    it('react-component-flow', () => {
-        const output = codehawk.analyzeProject(`${cwd}/samples/react-component-flow`)
-        assert.ok(output)
+    // it('react-component-flow', () => {
+    //     const output = codehawk.analyzeProject(`${cwd}/samples/react-component-flow`)
+    //     assert.ok(output)
 
-        const expectedRaw = fs.readFileSync(`${cwd}/samples/react-component-flow/expected.json`)
-        const expected = JSON.parse(expectedRaw)
+    //     const expectedRaw = fs.readFileSync(`${cwd}/samples/react-component-flow/expected.json`)
+    //     const expected = JSON.parse(expectedRaw)
 
-        assert.deepEqual(output.results, expected.results)
-    })
+    //     assert.deepEqual(output.results, expected.results)
+    // })
 
-    it('simple-class', () => {
-        const output = codehawk.analyzeProject(`${cwd}/samples/simple-class`)
-        assert.ok(output)
+    // it('simple-class', () => {
+    //     const output = codehawk.analyzeProject(`${cwd}/samples/simple-class`)
+    //     assert.ok(output)
 
-        const expectedRaw = fs.readFileSync(`${cwd}/samples/simple-class/expected.json`)
-        const expected = JSON.parse(expectedRaw)
+    //     const expectedRaw = fs.readFileSync(`${cwd}/samples/simple-class/expected.json`)
+    //     const expected = JSON.parse(expectedRaw)
 
-        assert.deepEqual(output.results, expected.results)
-    })
+    //     assert.deepEqual(output.results, expected.results)
+    // })
 
-    it('simple-es6-imports', () => {
-        const output = codehawk.analyzeProject(`${cwd}/samples/simple-es6-imports`)
-        assert.ok(output)
+    // it('simple-es6-imports', () => {
+    //     const output = codehawk.analyzeProject(`${cwd}/samples/simple-es6-imports`)
+    //     assert.ok(output)
 
-        const expectedRaw = fs.readFileSync(`${cwd}/samples/simple-es6-imports/expected.json`)
-        const expected = JSON.parse(expectedRaw)
+    //     const expectedRaw = fs.readFileSync(`${cwd}/samples/simple-es6-imports/expected.json`)
+    //     const expected = JSON.parse(expectedRaw)
 
-        assert.deepEqual(output.results, expected.results)
-    })
+    //     assert.deepEqual(output.results, expected.results)
+    // })
 
-    it('sweetalert', () => {
-        const output = codehawk.analyzeProject(`${cwd}/samples/sweetalert`)
-        assert.ok(output)
+    // it('sweetalert', () => {
+    //     const output = codehawk.analyzeProject(`${cwd}/samples/sweetalert`)
+    //     assert.ok(output)
 
-        const expectedRaw = fs.readFileSync(`${cwd}/samples/sweetalert/expected.json`)
-        const expected = JSON.parse(expectedRaw)
+    //     const expectedRaw = fs.readFileSync(`${cwd}/samples/sweetalert/expected.json`)
+    //     const expected = JSON.parse(expectedRaw)
 
-        assert.deepEqual(output.results, expected.results)
-    })
+    //     assert.deepEqual(output.results, expected.results)
+    // })
 })
